refactor(ckad-angular): create secondary entry points
