--- apps/frontend/src/app/modules/ckapps-design-system/ckad-ui-common/ckad-button/ckad-button.module.scss
@@ -1,1 +1,0 @@
-@import './scss/ckad-button.variables.scss';
