--- apps/frontend/src/app/modules/ckapps-design-system/ckad-ui-common/ckad-button/components/button/button.component.scss
@@ -1,35 +1,0 @@
-@import './button.variables.scss';
-
-@mixin ckad-button-variant($variant, $color) {
-  .ckad-button--color-#{$variant} {
-    background-color: $color;
-  }
-}
-
-@mixin ckad-button() {
-  box-sizing: border-box;
-  position: relative;
-  user-select: none;
-  cursor: pointer;
-  outline: 0;
-  border: none;
-  -webkit-tap-highlight-color: transparent;
-  display: inline-block;
-  white-space: nowrap;
-  text-decoration: none;
-  vertical-align: baseline;
-  text-align: center;
-  margin: 0;
-
-  // from variables
-  min-width: $ckad-button--min-width;
-  line-height: $ckad-button--line-height;
-  padding: $ckad-button--padding-x $ckad-button--padding-y;
-  border-radius: $ckad-button--border-radius;
-}
-
-:host {
-  // @include ckad-button();
-  // width: 100%;
-  // height: 100%;
-}
