--- apps/frontend/src/app/modules/ckapps-design-system/ckad-ui-common/ckad-button/components/button/button.variables.scss
@@ -1,5 +1,0 @@
-$ckad-button--border-radius: 2px !default;
-$ckad-button--line-height: 36px !default;
-$ckad-button--min-width: 90px !default;
-$ckad-button--padding-x: 0 !default;
-$ckad-button--padding-y: 16px !default;
