--- apps/frontend/src/app/modules/ckapps-design-system/ckad-ui-common/ckad-button/scss/ckad-button.variables.scss
@@ -1,1 +1,0 @@
-@import '../components/button/button.variables.scss';
