--- apps/frontend/src/app/modules/ckapps-design-system/ckad-ui-common/ckad-list/components/list/list.variables.scss
@@ -1,2 +1,0 @@
-// Variables for list
-$ckad-list--default--border-radius: 0.25rem !default;
