+++ apps/frontend/src/app/modules/ckapps-design-system/ckad-ui-common/scss/ckad-ui-common.submodules.scss
@@ -1,3 +1,2 @@
 // Import submodules
-@import '../ckad-button/ckad-button.module.scss';
 @import '../ckad-input/ckad-input.module.scss';
