+++ apps/frontend/src/app/modules/domain/recipe/modules/recipe-input/recipe-input.module.ts
@@ -6,9 +6,9 @@ import { MatAutocompleteModule } from '@angular/material/autocomplete';
 import { MatFormFieldModule } from '@angular/material/form-field';
 import { MatInputModule } from '@angular/material/input';
 import { MatListModule } from '@angular/material/list';
-import { CkadButtonModule } from '../../../../ckapps-design-system/ckad-ui-common/ckad-button/ckad-button.module';
+import { CkadButtonModule } from '@ckapp/angular/button';
+import { CkadListModule } from '@ckapp/angular/list';
 import { CkadInputModule } from '../../../../ckapps-design-system/ckad-ui-common/ckad-input/ckad-input.module';
-import { CkadListModule } from '../../../../ckapps-design-system/ckad-ui-common/ckad-list/ckad-list.module';
 import { CkadUiCommonModule } from '../../../../ckapps-design-system/ckad-ui-common/ckad-ui-common.module';
 import { IngredientModule } from '../../../ingredient/ingredient.module';
 import { RecipeInputSourceComponent } from './components/recipe-input-source/recipe-input-source.component';
