+++ libs/@ckapp/angular/button/src/lib/button-base.type.ts
