+++ libs/@ckapp/angular/list/src/lib/list-item.component.html
