+++ libs/@ckapp/angular/list/src/lib/list-item.component.spec.ts
