+++ libs/@ckapp/angular/list/src/lib/list.component.ts
