+++ libs/@ckapp/angular/list/src/lib/mutable-list-item.component.html
