+++ libs/@ckapp/angular/list/src/lib/mutable-list.component.html
