+++ libs/@ckapp/angular/list/src/lib/mutable-list.component.ts
@@ -1,6 +1,6 @@
 import { AfterContentInit, Component, contentChildren } from '@angular/core';
-import { CkadMutableListItemComponent } from '../mutable-list-item/mutable-list-item.component';
-import { CkadListComponent } from '../list/list.component';
+import { CkadListComponent } from './list.component';
+import { CkadMutableListItemComponent } from './mutable-list-item.component';
 
 /**
  * **ckapps design/component**
@@ -10,7 +10,6 @@ import { CkadListComponent } from '../list/list.component';
 @Component({
   selector: 'ckad-mutable-list',
   templateUrl: './mutable-list.component.html',
-  styleUrls: ['./mutable-list.component.scss'],
   imports: [CkadListComponent],
 })
 export class CkadMutableListComponent implements AfterContentInit {
